requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
django-modeltranslation==0.18.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
